Edit folders name, extention function, splite.getOrElse and remove dommy data
