+++ app/src/main/java/ir/saharapps/rickandmorty/data/di/DatabaseModule.kt
@@ -7,16 +7,15 @@ import dagger.Provides
 import dagger.hilt.InstallIn
 import dagger.hilt.android.qualifiers.ApplicationContext
 import dagger.hilt.components.SingletonComponent
-import ir.saharapps.rickandmorty.data.local.CharacterDatabase
 import ir.saharapps.rickandmorty.utils.Constants
 import javax.inject.Singleton
 
-@Module
-@InstallIn(SingletonComponent::class)
-object DatabaseModule {
-
-    @Singleton
-    @Provides
-    fun provideDataBase(@ApplicationContext context: Context) =
-        Room.databaseBuilder(context, CharacterDatabase::class.java, Constants.DATABASE_NAME).build()
-}
+//@Module
+//@InstallIn(SingletonComponent::class)
+//object DatabaseModule {
+//
+//    @Singleton
+//    @Provides
+//    fun provideDataBase(@ApplicationContext context: Context) =
+//        Room.databaseBuilder(context, CharacterDatabase::class.java, Constants.DATABASE_NAME).build()
+//}
