+++ app/src/main/java/ir/saharapps/rickandmorty/data/local/CharacterDao.kt
@@ -6,15 +6,15 @@ import androidx.room.OnConflictStrategy
 import androidx.room.Query
 import ir.saharapps.rickandmorty.data.dto.Character
 
-@Dao
-interface CharacterDao {
-
-    @Insert(onConflict = OnConflictStrategy.REPLACE)
-    suspend fun addCharacter(character: Character)
-
-    @Query("SELECT * FROM character")
-    suspend fun getAllCharacters(): List<Character>
-
-    @Query("SELECT * FROM  character WHERE id = :id")
-    suspend fun getCharacterById(id: Int): Character
-}
+//@Dao
+//interface CharacterDao {
+//
+//    @Insert(onConflict = OnConflictStrategy.REPLACE)
+//    suspend fun addCharacter(character: Character)
+//
+//    @Query("SELECT * FROM character")
+//    suspend fun getAllCharacters(): List<Character>
+//
+//    @Query("SELECT * FROM  character WHERE id = :id")
+//    suspend fun getCharacterById(id: Int): Character
+//}
