+++ app/src/main/java/ir/saharapps/rickandmorty/data/local/CharacterDatabase.kt
@@ -4,7 +4,7 @@ import androidx.room.Database
 import androidx.room.RoomDatabase
 import ir.saharapps.rickandmorty.data.dto.Character
 
-@Database(entities = [Character::class], version = 1)
-abstract class CharacterDatabase: RoomDatabase() {
-    abstract fun characterDao(): CharacterDao
-}
+//@Database(entities = [Character::class], version = 1)
+//abstract class CharacterDatabase: RoomDatabase() {
+//    abstract fun characterDao(): CharacterDao
+//}
