+++ app/src/main/java/ir/saharapps/rickandmorty/data/repository/CharactersRemoteRepositoryImpl.kt
@@ -1,20 +1,18 @@
 package ir.saharapps.rickandmorty.data.repository
 
 import ir.saharapps.rickandmorty.data.dto.Character
-import ir.saharapps.rickandmorty.data.local.CharacterDatabase
 import ir.saharapps.rickandmorty.data.remote.RickAndMortyApi
 import javax.inject.Inject
 
 class CharactersRemoteRepositoryImpl @Inject constructor(
     private val api: RickAndMortyApi,
-    private val database: CharacterDatabase
 ): CharactersRepository {
 
     override suspend fun getCharacters(): List<Character> {
         val characterList = api.getCharacters().characters
-        for(character in characterList){
-            database.characterDao().addCharacter(character)
-        }
+//        for(character in characterList){
+//            database.characterDao().addCharacter(character)
+//        }
         return characterList
     }
