+++ app/src/main/java/ir/saharapps/rickandmorty/domain/model/DetailCharacterViewState.kt
@@ -3,7 +3,7 @@ package ir.saharapps.rickandmorty.domain.model
 
 data class DetailCharacterViewState(
     val viewState: ViewState = ViewState.INITIAL,
-    val detailCharacter: DetailCharacter = DetailCharacter(1,"","","","","", emptyList()),
+    val detailCharacter: DetailCharacter? = null,
     val episodeList: List<Episode> = emptyList()
 )
