+++ app/src/main/java/ir/saharapps/rickandmorty/ui/screen/characters/CharactersViewModel.kt
@@ -1,4 +1,4 @@
-package ir.saharapps.rickandmorty.ui.screen.characters_screen
+package ir.saharapps.rickandmorty.ui.screen.characters
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -6,6 +6,7 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import ir.saharapps.rickandmorty.domain.model.CharacterViewState
 import ir.saharapps.rickandmorty.domain.model.ViewState
 import ir.saharapps.rickandmorty.domain.usecase.CharactersUseCase
+import ir.saharapps.rickandmorty.ui.screen.utility.updateState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
@@ -22,25 +23,17 @@ class CharactersViewModel @Inject constructor(
     val viewState: StateFlow<CharacterViewState> = _viewState.asStateFlow()
     fun getCharacters(){
         viewModelScope.launch(Dispatchers.IO) {
-            _viewState.value = CharacterViewState().update { copy(viewState = ViewState.LOADING) }
+            _viewState.updateState { copy(viewState = ViewState.LOADING) }
             try {
                 val characters = characterUseCase.getAllCharacters()
-                _viewState.value = _viewState.value.update {
+                _viewState.updateState {
                     copy(viewState = ViewState.SUCCESS, characters = characters)
                 }
             }catch (e: Exception){
-                _viewState.value = _viewState.value.update{copy(viewState = ViewState.FAILED)}
+                _viewState.updateState{copy(viewState = ViewState.FAILED)}
             }
         }
     }
 
 }
-
-fun <T> T.update(newState: T.() -> T): T{
-    return newState()
-}
-
-//fun <T> T.updateState(newState: T.() -> T) {
-//    val value = newState
-//}
