+++ app/src/main/java/ir/saharapps/rickandmorty/ui/screen/detailcharacter/DetailCharacterViewModel.kt
@@ -1,4 +1,4 @@
-package ir.saharapps.rickandmorty.ui.screen.detail_character
+package ir.saharapps.rickandmorty.ui.screen.detailcharacter
 
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
@@ -6,13 +6,12 @@ import dagger.hilt.android.lifecycle.HiltViewModel
 import ir.saharapps.rickandmorty.domain.model.DetailCharacterViewState
 import ir.saharapps.rickandmorty.domain.model.ViewState
 import ir.saharapps.rickandmorty.domain.usecase.CharactersUseCase
+import ir.saharapps.rickandmorty.ui.screen.utility.updateState
 import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.flow.MutableStateFlow
 import kotlinx.coroutines.flow.StateFlow
 import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.launch
-import java.net.ConnectException
-import java.net.SocketTimeoutException
 import javax.inject.Inject
 
 @HiltViewModel
@@ -24,25 +23,19 @@ class DetailCharacterViewModel @Inject constructor(
     val viewState: StateFlow<DetailCharacterViewState> = _viewState.asStateFlow()
 
     fun getCharacterDetail(characterId: Int){
+
         viewModelScope.launch(Dispatchers.IO) {
-            _viewState.value = DetailCharacterViewState().update { copy(viewState = ViewState.LOADING) }
+            _viewState.updateState { copy(viewState = ViewState.LOADING) }
             try {
                 val characterInfo = useCase.getCharacterById(characterId)
                 val episode = useCase.getEpisodeList(characterId)
-                _viewState.value = DetailCharacterViewState().update {
+                _viewState.updateState {
                     copy(viewState = ViewState.SUCCESS, detailCharacter = characterInfo, episodeList = episode)
                 }
             }catch (e: Exception){
-                _viewState.value = DetailCharacterViewState().update { copy(viewState = ViewState.FAILED) }
+                _viewState.updateState { copy(viewState = ViewState.FAILED) }
             }
         }
+
     }
 }
-
-fun <T> T.update(newState: T.() -> T): T{
-    return newState()
-}
-
-//fun <T> T.updateState(newState: T.() -> T) {
-//    val value = newState
-//}
