+++ app/src/main/java/ir/saharapps/rickandmorty/ui/screen/utility/UpdateViewStateExt.kt
@@ -1,0 +1,7 @@
+package ir.saharapps.rickandmorty.ui.screen.utility
+
+import kotlinx.coroutines.flow.MutableStateFlow
+
+fun <T> MutableStateFlow<T>.updateState(newState: T.() -> T) {
+    value = newState(value)
+}
