+++ app/src/main/res/layout/fragment_characters.xml
@@ -4,7 +4,7 @@
     android:layout_height="match_parent"
     xmlns:app="http://schemas.android.com/apk/res-auto"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context=".ui.screen.characters_screen.CharactersFragment">
+    tools:context=".ui.screen.characters.CharactersFragment">
 
     <ProgressBar
         android:id="@+id/pb_loading"
