+++ app/src/main/res/layout/fragment_detail_character.xml
@@ -4,7 +4,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="ui.screen.detail_character.DetailCharacterFragment">
+    tools:context="ui.screen.detailcharacter.DetailCharacterFragment">
 
     <ProgressBar
         android:id="@+id/pb_loading"
